Further @mattklein123 review feedback.

+++ include/envoy/network/connection.h
@@ -151,7 +151,7 @@ public:
    * Set a soft limit on the size of the read buffer prior to flushing to further stages in the
    * processing pipeline.
    */
-  virtual void setReadBufferLimit(size_t limit) PURE;
+  virtual void setReadBufferLimit(uint32_t limit) PURE;
 };
 
 typedef std::unique_ptr<Connection> ConnectionPtr;
