+++ test/common/network/connection_impl_test.cc
@@ -131,7 +131,7 @@ TEST(ConnectionImplTest, BufferStats) {
 
 class ReadBufferLimitTest : public testing::Test {
 public:
-  void readBufferLimitTest(uint32_t read_buffer_limit, size_t expected_chunk_size) {
+  void readBufferLimitTest(uint32_t read_buffer_limit, uint32_t expected_chunk_size) {
     const uint32_t buffer_size = 256 * 1024;
 
     Stats::IsolatedStoreImpl stats_store;
@@ -159,7 +159,7 @@ public:
           EXPECT_EQ("", server_connection->nextProtocol());
         }));
 
-    size_t filter_seen = 0;
+    uint32_t filter_seen = 0;
 
     EXPECT_CALL(*read_filter, onNewConnection());
     EXPECT_CALL(*read_filter, onData(_))
